Renew the SSL key using a local certificate authority
The previously used SSL key was expired and could not easily be renewed
because it was signed using a real CA (TERENA). To prevent having to
deal with this problem every year, this commit changes the trade-off
between security and ease of setting up the VM in the following manner:

 - use a local CA for the development VM so everything works
   out-of-the-box
 - use a SSL certificate with a long expiration period
 - bundle the SSL keys in the repository so there is no prerequisite
   action regarding SSL when provisioning the VM
 - on the other side of the trade-off: developers must trust the local
   CA once in their browser

The ssl/README.md file contains more information on the new SSL setup.