Add ra.feature behat feature
This test will contain the RA specific test cases we like to add to our
smoke tests. Two scenarios are provided:

1. The promotion of an indentity to become RA.
2. The demotion of an RA to a regular user.
Upgrade Symfony to 2.7.49

To harden against CVE-2018-14773. In the process Twig was also upgraded
to satisfy Symfony's 2.7.49 requirements.
Remodel the StepUpAuthenticationService

Some of the application logic was moved to the API, this to be able to
mock the Api service. In essence the YubikeyOtpVerificationResult was
moved to the API, becoming the return type for said service.

The verification logic that was performed in the Gateway bundle was
moved to the ApiBundle.

This was the most elegant way to then provide a test double. An
alternative was to also provide a test double for the
StepUpAuthenticationService, which is a very godlike service which would
result in losing too much important business logic for the E2E tests.

See: https://www.pivotaltracker.com/story/show/156570267