Use the SMS Service test double in test
The test double is used to prevent the sending of actual SMS messages to
 a test user.

In order to make the change work correctly, the VettingService now works
with the SmsSecondFactorServiceInterface. This interface will be
introduced in the upcoming Stepup-bundle release.

And removed the die statement, preventing the tests to start running