Pin SSP to a specific version
SimpleSamlPHP is now pinned to revision:
adf1eb86215b3689a3c766dfa81cb9d55b786237. This is the last known version
that works with the DebugSP. Later releases might work fine, but are
not yet compatible with the behat tests and or the DebugSP module. Urls
seem to have been changed.