Provide profile page behat coverage
Two scenarios have been added testing the various features of the
profile page.

Scenario 1 tests the correct display of all possible values on the
profile page.

Scenario 2 tests the correct displaying of implicit roles on the profile
athoriszations section.

Other scenarios have been added for RA users and the correct determinaiton
of implicit roles for that type of role.