Enhance location test coverage to add and edit location
The behat test coverage is enhanced by testing the location logic on
the location page.

The tests for the removal of a location is omitted because of the
extensive js use. Therefore it will cost extra time to refactor the
code or change the used behat driver and this could be better addressed
when actual work is done on the locations themselves.