Enforce use_raa even if user is raa through select_raa
There was a bug which allowed RAA's to execute actions if
the user was made RAA through select_raa. There however still
needed to be validated if a user could act as RAA on a
certain institution by also enforcing use_raa.

Also a bug was fixed which incorrectly set the actorId in tests.
This was done through actor while actor_id should have been used.