Change SF4 default env based on dev group_names
By default, when the build is not in development mode, the production
debug=false option is used. When in dev mode, debug mode is engaged, and
a conditional smoketest env can be set if the testcookie was present.

Should be fine for production releases as they will not be polluted with
the additional config.

To pull this change run: './deploy-site-app.sh'

Should be rather compatible with the Stepup RV changes found here:
https://github.com/OpenConext/Stepup-Deploy/blob/572d09b8dc068710af3f39dcaccbca627293a9f7/roles/app/templates/php72-nginx-vhost-symfony4.conf.j2#L45-L51