Add event data integrity test
The Middleware command handler tests the correct data on an Event, this
however is the deserialized data that is constructed on the event. The
event payload may contain additional data which in turn could be
sensitive data.

This behat test is created to test the presence of certain data on the
event.