Provide Self Asserted Tokens test coverage
To get the test framework up and running, some tweaking of the test
suite was required. The behat tests now run on the dockerized stepup
deploy centos 7 container. So paths needed adjustment.

Some of the other features where updated to use the updated step
definitions. With this PR, the SAT tests should pass, so should the
other features touched in this commit. I can not guarantee all other
features also give a green light. There might be need for some
additional tweaking to the platform. But for this PR that was out of
scope.